WEB/YAML2JSON/yamlconverter/requirements.txt
Django==4.2.1
ruamel.yaml==0.17.17